병합 실패
